Use URLStorage.structure directly in setters and mutating views.
Previously, when there were multiple header types, we had to precalculate
the final structure as a URLStructure<Int>, as it supported all URLs.
That meant if the mutation needed to change the header type, it could just
copy the code-units to the new buffer and narrow the structure as needed.

Now that we only have a single header type, we have a single native structure
width and don't need to keep a function-local copy using a different integer type.
We can just refer to the existing structure in the header directly.

This opens the door to perhaps not even precalculating the final structure at all -
we could just calculate the new values of fields that might change,
replace the code-units, and modify the structure in-place. That means we wouldn't
need a function-local copy of the structure and wouldn't need to pass
the new structure around to helper functions, further reducing code-size
and perhaps giving even better performance.