Redo lazy percent-encoding.
Now uses a single type rather than FlattenSequence, since FS has poor performance
especially when subscripting the outer collection is expensive (like, errr... us).
This gives a solid 35%+ performance improvement, and writing the contents collapses
to a nice, tight loop with as few branches as I could manage.

There may be some ways to improve it a bit more, but right now I think the biggest
improvement would be if the compiler would emit the percent-encoding table in to the DATA section,
so we don't have to call the 'unsafeMutableAddressor'. Also, it could eliminate
the cmov used to bounds-check the table, or we could remove it.
I have some ideas, but they're funky.