Move bounds-checking during construction to the writing phase.
It's less safe and performant to do this in StructureAndMetricsCollector,
as it always operates without hints and so always sees non-contiguous, percent-encoded
pieces. The capacity and structure it calculates, whilst important for correctness,
are not crucial for memory safety. Maybe we allocate too much, or not enough,
or the Collection is invalid and we end up with a broken structure.
It doesn't matter, because...

Instead, bounds-checking is now part of UnsafePresizedBufferWriter.
That's the part which does the actual writing, so it's ^actually^ the place where
a malicious Collection could introduce a buffer overflow. Additionally,
when a piece does not require percent-encoding (which is reasonably often),
it will see contiguous slices of data straight from the source, so it's faster as well.