[Path parser] _DeferredPathComponent no longer stores a slice of the input string.
I noticed the compiler was generating a lot of code to store the index ranges,
but the parser almost never cares about the location of a deferred Windows drive,
only its contents (which are known to be 2 bytes).

This enables some significant refactoring and clean-up:
- _DeferredPathComponent no longer has to be generic
- It can now live inside _PathParserState without forcing that to be generic, too
- Helper functions which took _both_ an inout _DPC and _PPS can be simplified
- Visitor callbacks can't accept _DPCs content any more, because it's no longer
  a slice of the input collection - but that's actually good! Splitting these
  leads to better, faster code which is easier to understand.

I'm seeing ~5% improvement to AverageURLs and similar benchmarks,
and a 10% improvement on "Long Paths". More can be done, but this is a solid start.