[Path parser] Stop deferring empty components and simplify the parser.
Deferring empty components used to be necessary, because
older versions of the standard would discard consecutive empties
at the start of file URLs. But that isn't the case now,
and calculating the path sigil can be done much more easily.

This allows the parser to be dramatically simplified, resulting in
faster, smaller code which is easier to understand and maintain.